+++ spec/config_specs.clj
@@ -1,17 +1,33 @@
 (ns config-specs
   (:use io.aviso.config
         speclj.core)
-  (:require [schema.core :as s]))
+  (:require [schema.core :as s]
+            [com.stuartsierra.component :as component]))
 
-(s/defschema WebServer {:web-server {:port      s/Int
+(s/defschema WebServerConfig {:web-server {:port s/Int
                                      :pool-size s/Int}})
 
-(s/defschema Database {:database
+(s/defschema DatabaseConfig {:database
                        {:hostname s/Str
                         :user     s/Str
                         :password s/Str}})
 
 (s/defschema Env {:home s/Str})
+
+(defrecord WebServer [configuration port pool-size]
+
+  component/Lifecycle
+
+  (start [this]
+    (assoc this :port (get-in configuration [:web-server :port])))
+
+  (stop [this] this))
+
+(defn new-web-server
+  []
+  (-> (map->WebServer {})
+      (component/using [:configuration])
+      (with-config-schema WebServerConfig)))
 
 (describe "io.aviso.config"
 
@@ -35,34 +51,34 @@
 
   (it "can parse YAML"
       (->> (assemble-configuration {:prefix  "yaml"
-                                    :schemas [WebServer]})
+                                    :schemas [WebServerConfig]})
            ;; note: coercion to s/Int occurred
            (should= {:web-server {:port      8080
                                   :pool-size 50}})))
 
   (it "can parse EDN"
       (->> (assemble-configuration {:prefix  "edn"
-                                    :schemas [WebServer]})
+                                    :schemas [WebServerConfig]})
            (should= {:web-server {:port      8080
                                   :pool-size 25}})))
 
   (it "overrides default profile with other profiles"
       (->> (assemble-configuration {:prefix   "order1"
-                                    :schemas  [WebServer]
-                                    :profiles [:testing]})
+                                    :schemas  [WebServerConfig]
+                                    :profiles [:web]})
            (should= {:web-server {:port      9090
                                   :pool-size 40}})))
 
   (it "overrides default profile with nil profile"
       (->> (assemble-configuration {:prefix   "order2"
-                                    :schemas  [WebServer]
+                                    :schemas  [WebServerConfig]
                                     :profiles [:testing]})
            (should= {:web-server {:port      9090
                                   :pool-size 40}})))
 
   (it "mixes together multiple profiles and schemas"
       (->> (assemble-configuration {:prefix           "mix"
-                                    :schemas          [WebServer Database]
+                                    :schemas          [WebServerConfig DatabaseConfig]
                                     :additional-files ["dev-resources/mix-production-overrides.yaml"]
                                     :overrides        {:web-server {:port 9999}}})
            (should= {:web-server {:port      9999
@@ -73,7 +89,7 @@
 
   (it "processes arguments"
       (->> (assemble-configuration {:prefix  "mix"
-                                    :schemas [WebServer Database]
+                                    :schemas [WebServerConfig DatabaseConfig]
                                     :args    ["--load" "dev-resources/mix-production-overrides.yaml"
                                               "web-server/port=9999"
                                               "database/hostname=db"]})
@@ -83,41 +99,51 @@
                                   :user     "prod"
                                   :password "secret"}})))
 
-  (it "expands environment variables"
+  (context "expansions"
+    (it "expands environment variables"
 
-      (->> (assemble-configuration {:prefix  "env"
-                                    :schemas [Env]})
-           (should= {:home @home})))
+        (->> (assemble-configuration {:prefix  "env"
+                                      :schemas [Env]})
+             (should= {:home @home})))
 
-  (it "expands environment variables on edn files"
-      (->> (assemble-configuration {:prefix "envedn"
-                                    :schemas [Env]})
-           (should= {:home @home})))
+    (it "expands environment variables on edn files"
+        (->> (assemble-configuration {:prefix  "envedn"
+                                      :schemas [Env]})
+             (should= {:home @home})))
 
-  (it "can use a default value on an environment variable"
-      (->> (assemble-configuration {:prefix "envdef"
-                                    :schemas [{s/Any s/Any}]})
-           (should= {:use-default "default-plugh"
-                     :use-env     @home})))
+    (it "can use a default value on an environment variable"
+        (->> (assemble-configuration {:prefix  "envdef"
+                                      :schemas [{s/Any s/Any}]})
+             (should= {:use-default "default-plugh"
+                       :use-env     @home})))
 
-  (it "can expand non-environment variable properties"
-      (->> (assemble-configuration {:prefix     "vars"
-                                    :properties {:special "an-option"}
-                                    :schemas    [{s/Any s/Any}]})
-           (should= {:unmatched "ok"
-                     :special   "an-option"
-                     :home      @home})))
+    (it "can expand non-environment variable properties"
+        (->> (assemble-configuration {:prefix     "vars"
+                                      :properties {:special "an-option"}
+                                      :schemas    [{s/Any s/Any}]})
+             (should= {:unmatched "ok"
+                       :special   "an-option"
+                       :home      @home})))
 
-  (it "can expand JVM system properties"
-      (->> (assemble-configuration {:prefix  "sysprops"
-                                    :schemas [{s/Any s/Any}]})
-           (should= {:user-home (System/getProperty "user.home")})))
+    (it "can expand JVM system properties"
+        (->> (assemble-configuration {:prefix  "sysprops"
+                                      :schemas [{s/Any s/Any}]})
+             (should= {:user-home (System/getProperty "user.home")}))))
 
   (it "can associate and extract schemas"
-      (->> [(with-config-schema {} WebServer)
+      (->> [(with-config-schema {} WebServerConfig)
             {}
-            (with-config-schema {} Database)]
+            (with-config-schema {} DatabaseConfig)]
            extract-schemas
-           (should= [WebServer Database]))))
+           (should= [WebServerConfig DatabaseConfig])))
+
+  (it "can build a system"
+      (should= 9999
+               (-> (component/system-map :web-server (new-web-server))
+                   (extend-system-map {:prefix   "system"
+                                       :profiles [:web-server]})
+                   component/start-system
+                   :web-server
+                   :port))))
 
 (run-specs)